global: move show-log-messages
Move <show-log-message> from the Index to the Generic Menu.
Now it can be called from anywhere in NeoMutt.

Also, add a synonym <error-history> for Mutt compatibility.
